feat: improved multi plate slicing for 3mf files
