+++ .gitignore
@@ -136,3 +136,5 @@ dist
 .pnp.*
 
 .test/
+
+.DS_Store
