+++ package.json
@@ -13,6 +13,7 @@
   "license": "AGPL-3.0-or-later",
   "dependencies": {
     "@types/multer": "^1.4.12",
+    "archiver": "^7.0.1",
     "express": "^5.1.0",
     "multer": "^2.0.0",
     "swagger-ui-express": "^5.0.1"
@@ -20,8 +21,10 @@
   "type": "module",
   "devDependencies": {
     "@eslint/js": "^9.26.0",
+    "@types/archiver": "^6.0.3",
     "@types/express": "^5.0.1",
     "@types/node": "^22.15.17",
+    "@types/swagger-ui-express": "^4.1.8",
     "eslint": "^9.26.0",
     "globals": "^16.1.0",
     "tsc-esm-fix": "^3.1.2",
