+++ src/routes/slicing/models.ts
@@ -3,6 +3,12 @@ export interface SlicingSettings {
   preset: string;
   filament: string;
   bedType: string;
+  allPlates?: boolean;
+}
+
+export interface SliceResult {
+  gcodes: string[];
+  workdir: string;
 }
 
 export type Category = "printers" | "presets" | "filaments";
