--- src/swagger-ui-express.d.ts
@@ -1,1 +1,0 @@
-declare module 'swagger-ui-express';
